feat: getTextOffset 유틸리티 함수로 분리
